# at/fhtw/swkom/paperless/openapi.yaml
openapi: 3.1.0
info:
  title: Paperless REST-Server
  contact:
    name: Bernhard Wallisch
    url: https://www.technikum-wien.at/personal/bernhard-wallisch/
  version: 1.0.0
servers:
  - url: 'localhost:8081'
paths:
  /api/documents/post_document:
    post:
      summary: "Upload a document"
      operationId: postDocument
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: "Document successfully uploaded."
        400:
          description: "Document upload failed due to bad request!"
  /api/documents:
    get:
      summary: "Return a list of documents"
      operationId: getDocuments
      responses:
        200:
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/documents'
  /api/documents/{id}:
    get:
      summary: "Retrieve a document by id"
      operationId: getDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The id of the document
      responses:
        200:
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/document'
        404:
          description: "Document does not exist with this id!"
    put:
      summary: "Update a document by id"
      operationId: updateDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The id of the Document
      responses:
        200:
          description: "Document updated successfully"
        404:
          description: "Document does not exist with this id!"
    delete:
      summary: "Delete a document by id"
      operationId: deleteDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The id of the document
      responses:
        204:
          description: "Document deleted successfully"
        404:
          description: "Document does not exist with this id!"
components:
  schemas:
    document:
      type: object
      properties:
        id:
          type: integer
          description: "The id of the document"
        title:
          type: string
          description: "The title of the document"
        author:
          type: string
          description: "The author of the document's contents"
        created:
          type: string
      required:
        - title
    documents:
      type: array
      items:
        $ref: '#/components/schemas/document'